scripts: allow adding hardware interfaces to bonding interfaces
On platforms that support adding hardware interfaces to bonding
interfaces (and only hardware interfaces), as isgnalled by the
bonding.hardware-members-only platform capability, then allow hardware
interfaces (and only hardware interfaces) to be added as members of
the bonding interface.

Adding the restriction that on such platforms software-only bonding
interfaces are not supported then it simplifies the implementation and
there is no real use for software bonding on a hardware platform.

VRVDR-49261