yang: add support for configuring storm-control
For bonding interfaces created in hardware there is an expectation
that the user can configure the same features on them as for physical
front-panel interfaces. Storm control is available on such interfaces
on some platforms, so add support for configuring storm control on
bonding interfaces by creating a new YANG file and package for it that
can then be pulled into the image through package lists.

VRVDR-49261