yang: add deviations file for hardware platforms
Due to resource limitations in our system architecture (use of DPDK
port IDs) and in the hardware then we cannot support an unlimited
number of bonding interfaces, so set Broadcom StrataDNX platform
restrictions (including the platform used for SIAD, and the UfiSpace
Apollo NCP1-1) to a reasonable value of 32. The former is also true of
software interfaces, but due to YANG backwards compatibility
restrictions we can't change this now.

In addition, due to hardware limitations then we also cannot support
unlimited members to a bundle (although in practice this will be
limited to the number of front-panel ports anyway), so we set a limit
of 32 here which is a reasonable trade-off versus number of bonding
interfaces and allows plenty of scope for relaxation of either this or
the number of bonding interfaces in the future, should it be required.

VRVDR-49261