yang: add support for minimum-links configuration
There is a requirement to support a minimum-links configuration
option. This option prevents a bonding interface from coming up and
being usable for routing when it doesn't have enough members up to
give sufficient bandwidth for the expected traffic streams and to
allow the traffic to follow an alternative path which does have
sufficient bandwidth.

Make use of the existing libteam support for this where it the
proprerty is called min-ports. Since this is only supported in libteam
in LACP mode, and would be hard to make work properly in balanced
mode, then this option is restricted to LACP mode only.

VRVDR-49339