scripts: fix race between adding member and setting primary interface
Fix a race between adding a member to the bonding interface and
setting it as the primary interface when the member uses deferred
config application. In this case, there is no guarantee of ordering
since the two pieces of configuration are on different interfaces,
with the primary interface configuration rooted on the bonding
interface, and the bond-group configuration rooted on the member
interface.

Avoid this race by detecting when the primary configuration would be
prematurely applied and instead defer it until the member is actually
added.

Fixes: VRVDR-50025