Stop using vplane link_watch
The vplane link_watch was a good idea but it creates more problems than
it solves. The controller races against the kernel to send state updates
to the libteam. Since the state of the kernel bond is what's important
here, it doesn't make sense to confuse libteam or attempt to constantly
reconcile the differing state information.

The reason for vplane link_watch is our failure to correctly manage the
carrier (LOWER_UP) state of the tun interfaces. However, it looks like
the team driver generates its link events based on netif_oper_up().
This is probably more desireable since LOWER_UP is not sufficient in
all cases to indicate a link is "ready."

Bugfix: VRVDR-51623