vyatta-bonding: set admin state on H/W members correctly after changes
For hardware bonding members, setting the admin state of the member
interfaces independently of the bonding interface is
supported. However, adding/removing the member implicitly changes the
state. Therefore, fix up the admin state of the hardware interface
after adding/removing as a member such that it matches the configured
state.

Fixes: VRVDR-52550