vyatta-bonding: fix race between removing port and deleting bonding interface
When interface deferred config application is supported, there is a
race between the bonding interface being removed and removing a member
interface since they are applied asynchronously with each
other. Therefore, remove the validation that the bonding interface
must exist in remove_port and only remove as a member when the bonding
interface does still exist. This avoids an error being logged and
allows the admin-state processing to still be performed.

Fixes: VRVDR-50456