Add more efficient VIF validation to bonding interfaces.
The current per-VIF must statements for VIF validation are better than the previous script, but for 100s of VIFs on an interface they are still slow (many seconds). An XPATH plugin has now been written that allows us to validate the whole interface with one configd:must statement. If the plugin is present, the per-VIF must statement is not run, falling through to a configd:must 'true()' call, and the change here has the opposite logic where if the plugin is present, we run validation via configd:must instead of the 'true()' in the must statement.

Verified by acceptance test.

Part of VRVDR-53594.